Add rich OGP images for individual manhole pages
- Convert /manhole/[id] to Server Component wrapper so SNS crawlers
  receive og:image meta tags (previously all client-side DOM manipulation)
- Generate 1200x630 card OGP images via Sharp + SVG with photo, prefecture
  badge, city name, pokemon names, and title badge
- Support ?photo={photoId} query param: validates the photo belongs to the
  manhole and is public, then uses it as the card photo
- Fallback chain: specified photo → first public photo → SVG text card → default
